# perf: improve mem sync