--- a/config/config.exs
+++ b/config/config.exs
@@ -13,7 +13,8 @@ config :ae_plugin,
 
 config :ae_mdw,
   sync: true,
-  contract_cache_expiration_minutes: 1440
+  contract_cache_expiration_minutes: 60,
+  max_mem_sync_secs: 120
 
 # Endpoints
 config :ae_mdw, AeMdwWeb.Endpoint,