--- a/config/config.exs
+++ b/config/config.exs
@@ -13,8 +13,7 @@ config :ae_plugin,
 
 config :ae_mdw,
   sync: true,
-  contract_cache_expiration_minutes: 60,
-  max_mem_sync_secs: 120
+  contract_cache_expiration_minutes: 60
 
 # Endpoints
 config :ae_mdw, AeMdwWeb.Endpoint,