--- a/lib/ae_mdw/application.ex
+++ b/lib/ae_mdw/application.ex
@@ -10,7 +10,6 @@ defmodule AeMdw.Application do
   """
   alias AeMdw.Contract
   alias AeMdw.Db.Model
-  alias AeMdw.Db.State
   alias AeMdw.EtsCache
   alias AeMdw.Extract
   alias AeMdw.NodeHelper
@@ -228,8 +227,6 @@ defmodule AeMdw.Application do
   end
 
   def start_phase(:start_sync, _start_type, []) do
-    :ok = State.init_mem_state()
-
     if Application.fetch_env!(:ae_mdw, :sync) do
       Watcher.start_sync()
     end