--- a/lib/ae_mdw/sync/server.ex
+++ b/lib/ae_mdw/sync/server.ex
@@ -52,7 +52,7 @@ defmodule AeMdw.Sync.Server do
            | :idle
            | :stopped
            | {:syncing_db, reference()}
-           | {:syncing_mem, reference(), State.t()}
+           | {:syncing_mem, reference()}
   @typep state_data() :: %__MODULE__{
            chain_height: height(),
            chain_hash: hash(),
@@ -153,13 +153,9 @@ defmodule AeMdw.Sync.Server do
         {:next_state, {:syncing_db, ref}, state_data}
 
       db_height >= max_db_height and mem_hash != chain_hash ->
-        old_mem_state = State.mem_state()
+        ref = spawn_mem_sync(db_height, chain_hash)
 
-        ref =
-          State.create_mem_state()
-          |> spawn_mem_sync(db_height, chain_hash)
-
-        {:next_state, {:syncing_mem, ref, old_mem_state}, state_data}
+        {:next_state, {:syncing_mem, ref}, state_data}
 
       true ->
         :keep_state_and_data
@@ -180,9 +176,8 @@ defmodule AeMdw.Sync.Server do
     {:next_state, :idle, new_state_data, @internal_check_sync}
   end
 
-  def handle_event(:info, {ref, mem_hash}, {:syncing_mem, ref, old_mem_state}, state_data) do
+  def handle_event(:info, {ref, mem_hash}, {:syncing_mem, ref}, state_data) do
     new_state_data = %__MODULE__{state_data | mem_hash: mem_hash}
-    :ok = State.delete_mem_state(old_mem_state)
     {:next_state, :idle, new_state_data, @internal_check_sync}
   end
 
@@ -260,8 +255,9 @@ defmodule AeMdw.Sync.Server do
     end)
   end
 
-  defp spawn_mem_sync(mem_state, from_height, last_hash) do
+  defp spawn_mem_sync(from_height, last_hash) do
     spawn_task(fn ->
+      mem_state = State.create_mem_state()
       from_txi = Block.next_txi(mem_state)
 
       from_mbi =