--- a/test/ae_mdw_web/controllers/contract_controller_test.exs
+++ b/test/ae_mdw_web/controllers/contract_controller_test.exs
@@ -1,5 +1,5 @@
 defmodule AeMdwWeb.Controllers.ContractControllerTest do
-  use AeMdwWeb.ConnCase, async: false
+  use AeMdwWeb.ConnCase
   @moduletag skip_store: true
 
   alias :aeser_api_encoder, as: Enc