## What

Replace `GbTree` with ets based `SortedTable`.

## Why

refs #1389 + out of sync issue on mainnet on 5k+ contract calls

## Notes

Benchmark result for 10k contract calls with 2 event logs each:
```
root@3e66f18de36a:/app# MIX_ENV=test elixir -S mix run --no-start test/support/bench/sorted_table.exs 

Operating System: Linux
CPU Information: AMD Ryzen 7 5800H with Radeon Graphics
Number of Available Cores: 16
Available memory: 15.48 GB
Elixir 1.13.4
Erlang 23.3.4.18

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
memory time: 0 ns
parallel: 1
inputs: none specified
Estimated total run time: 14 s

Benchmarking GbTree...
Benchmarking SortedTable...

Name                  ips        average  deviation         median         99th %
SortedTable         12.77    0.00130 min     ±5.58%    0.00131 min    0.00149 min
GbTree             0.0148       1.13 min     ±0.00%       1.13 min       1.13 min

Comparison: 
SortedTable         12.77
GbTree             0.0148 - 863.34x slower +1.13 min
```